Improve dox and rename directory EmbeddingMaps to CoordinateMaps
