+++ src/Domain/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-add_subdirectory(EmbeddingMaps)
+add_subdirectory(CoordinateMaps)
 
 set(LIBRARY Domain)
