+++ src/Domain/CoordinateMaps/AffineMap.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Domain/EmbeddingMaps/AffineMap.hpp"
+#include "Domain/CoordinateMaps/AffineMap.hpp"
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
